+++ src/main/java/com/breakinblocks/plonk/Plonk.java
@@ -67,5 +67,6 @@ public class Plonk {
         RegistryTESRs.init();
         RenderTypeLookup.setRenderLayer(RegistryBlocks.placed_items, RenderType.getCutout());
         ClientEvents.registerKeyBindings();
+        // TODO REGISTER CLIENTSIDE COMMAND
     }
 }