+++ src/main/java/com/breakinblocks/plonk/client/render/tile/TESRPlacedItems.java
@@ -1,5 +1,6 @@
 package com.breakinblocks.plonk.client.render.tile;
 
+import com.breakinblocks.plonk.client.util.RenderUtils;
 import com.breakinblocks.plonk.common.block.BlockPlacedItems;
 import com.breakinblocks.plonk.common.tile.TilePlacedItems;
 import com.breakinblocks.plonk.common.tile.TilePlacedItems.ItemMeta;
@@ -23,7 +24,6 @@ import net.minecraft.util.Direction;
 import net.minecraft.util.math.vector.Matrix4f;
 import net.minecraft.util.math.vector.Quaternion;
 import net.minecraft.util.math.vector.Vector3f;
-import net.minecraftforge.common.model.TransformationHelper;
 
 import static com.breakinblocks.plonk.common.tile.TilePlacedItems.RENDER_TYPE_BLOCK;
 import static com.breakinblocks.plonk.common.tile.TilePlacedItems.RENDER_TYPE_ITEM;
@@ -37,6 +37,7 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
     private static final Tessellator tessellator = Tessellator.getInstance();
     private static final BufferBuilder bufferbuilder = tessellator.getBuffer();
     private static final BlockRendererDispatcher blockRendererDispatcher = mc.getBlockRendererDispatcher();
+    private static final double EPS = 0.001;
 
     public TESRPlacedItems(TileEntityRendererDispatcher rendererDispatcherIn) {
         super(rendererDispatcherIn);
@@ -44,8 +45,8 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
 
     public static int getRenderTypeFromStack(ItemStack itemstack) {
         IBakedModel model = itemRenderer.getItemModelWithOverrides(itemstack, null, null);
-        Matrix4f matrixFixed = TransformationHelper.toTransformation(model.getItemCameraTransforms().getTransform(ItemCameraTransforms.TransformType.FIXED)).getMatrix();
-        Matrix4f matrixGui = TransformationHelper.toTransformation(model.getItemCameraTransforms().getTransform(ItemCameraTransforms.TransformType.GUI)).getMatrix();
+        Matrix4f matrixFixed = RenderUtils.getModelTransformMatrix(model, ItemCameraTransforms.TransformType.FIXED);
+        Matrix4f matrixGui = RenderUtils.getModelTransformMatrix(model, ItemCameraTransforms.TransformType.GUI);
         Matrix4f difference = MatrixUtils.difference(matrixFixed, matrixGui);
         MatrixUtils.TransformData transform = new MatrixUtils.TransformData(difference);
 
@@ -63,8 +64,12 @@ public class TESRPlacedItems extends TileEntityRenderer<TilePlacedItems> {
         //String message = String.format("hS=%.3f hRot=%.3f\n", hS, hRot) + transform.toString();
         //for (String line : message.split("\r?\n"))
         //    Minecraft.getInstance().ingameGUI.getChatGUI().printChatMessage(new StringTextComponent(line));
-        // Change in scaling is < 1 and change in rotation at non-right angles;
-        return hS < (1.0 - 0.001) && hRot > 0.001 ? RENDER_TYPE_BLOCK : RENDER_TYPE_ITEM;
+        // The following is a heuristic
+        // Use block rendering: hRot ~= 83.80586
+        final double blockRot = 83.80586;
+        if (blockRot - EPS <= hRot && hRot <= blockRot + EPS)
+            return RENDER_TYPE_BLOCK;
+        return RENDER_TYPE_ITEM;
     }
 
     /**